+++ package.json
@@ -44,7 +44,7 @@
     "sinon-chai": "4.0.1",
     "source-map-support": "0.5.21",
     "typescript": "npm:@typescript/typescript6@6.0.1",
-    "typescript7": "npm:typescript@7.0.1-rc",
+    "@typescript/native": "npm:typescript@7.0.2",
     "typescript-eslint": "8.58.2"
   },
   "scripts": {
