+++ packages/typescript-checker/package.json
@@ -52,7 +52,7 @@
     "@stryker-mutator/test-helpers": "workspace:*",
     "@types/semver": "7.7.1",
     "tslib": "~2.8.0",
-    "typescript7": "npm:typescript@7.0.1-rc"
+    "@typescript/native": "npm:typescript@7.0.2"
   },
   "peerDependencies": {
     "@stryker-mutator/core": "workspace:*",
