+++ packages/typescript-checker/src/ts-native/text-helpers.ts
