+++ packages/typescript-checker/test/unit/create.spec.ts
@@ -3,7 +3,7 @@ import { expect } from 'chai';
 
 import { createTypescriptChecker } from '../../src/index.js';
 import { TypescriptChecker } from '../../src/typescript-checker.js';
-import { NativeTypescriptChecker } from '../../src/ts-v7/native-typescript-checker.js';
+import { NativeTypescriptChecker } from '../../src/ts-native/native-typescript-checker.js';
 import { TypescriptCheckerOptionsWithStrykerOptions } from '../../src/typescript-checker-options-with-stryker-options.js';
 
 describe(createTypescriptChecker.name, () => {
